Move generic types to their own module
