+++ src/api/api.utils.ts
@@ -2,8 +2,9 @@ import { ParsedUrlQuery } from 'querystring';
 import { COMMA_SEPARATED_LIST_REGEX } from '../components/FluidTypeScaleCalculator/Form/GroupTypeScaleSteps/GroupTypeScaleSteps.constants.';
 import { DEFAULT_FONT_FAMILY, initialFormState } from '../constants';
 import { QueryParamKey } from '../types';
+import type { NarrowById } from '../types.generics';
 import { isNumber, throwIf } from '../utils';
-import { NarrowById, QueryParam, QueryParamConfig, ValidatedQueryParam } from './api.types';
+import type { QueryParam, QueryParamConfig, ValidatedQueryParam } from './api.types';
 
 /** Returns the query parameter config entry corresponding to the param that has the specified ID/key. */
 export const getQueryParam = <K extends QueryParam['id']>(
